Switch size tracking to a u64
Common use cases like file transfers shouldn't fail on 32 bit platforms.

+++ src/lib.rs
@@ -64,7 +64,7 @@ impl<I> PbIter<I> where
 {
     pub fn new(iter: I) -> Self {
         let size = iter.size_hint().0;
-        PbIter {iter: iter, progress_bar: ProgressBar::new(size)}
+        PbIter {iter: iter, progress_bar: ProgressBar::new(size as u64)}
     }
 }
